+++ src/data/rasterizer.c
@@ -5,7 +5,7 @@
 #include "msdfgen-c.h"
 #include <math.h>
 
-Rasterizer* lovrRasterizerInit(Rasterizer* rasterizer, Blob* blob, int size) {
+Rasterizer* lovrRasterizerInit(Rasterizer* rasterizer, Blob* blob, float size) {
   stbtt_fontinfo* font = &rasterizer->font;
   unsigned char* data = blob ? blob->data : VarelaRound_ttf;
   if (!stbtt_InitFont(font, data, stbtt_GetFontOffsetForIndex(data, 0))) {
